Add redirects and refactor other areas.
-Router class has redirects
-Make not found as final case in RequestHandler
-ResponseBuilder is one method takes a Status
-CobSpecRoutes represents this specific router
-Add Routes interface to define what routes might be
-Add Redirect to Status enum
-DetermineHandler now takes a uri as well

+++ src/main/java/javaserver/RequestHandler.java
@@ -7,7 +7,7 @@ public class RequestHandler {
 
     private Request request;
     private Response response;
-    private Router router = new Router();
+    private Router router = new Router(new CobSpecRoutes());
     private String method;
     private ParameterDecoder decoder = new ParameterDecoder();
 
@@ -22,23 +22,33 @@ public class RequestHandler {
             response.setBody(body);
         }
 
+        if (router.isRedirect(uri)) {
+            response = new ResponseBuilder().buildResponse(Status.Redirect);
+            String headers = router.getRedirectHeader(uri);
+            response.setHeaders(headers);
+        }
         return response;
     }
 
     public Handler determineHandler(Request request) {
         String method = request.getMethod();
-        switch (method) {
-          case "GET":
-              return new GetHandler();
-          case "POST":
-          case "PUT" :
-              return new PostPutHandler();
-          case "DELETE":
-              return new DeleteHandler();
-          case "OPTIONS":
-              return new OptionsHandler(router);
-          default:
-              return new ErrorHandler();
+        String uri = request.getUri();
+        if (router.uriIsAllowed(method, uri)) {
+            switch (method) {
+              case "GET":
+              case "HEAD":
+                  return new GetHandler();
+              case "POST":
+              case "PUT" :
+                  return new PostPutHandler();
+              case "DELETE":
+                  return new DeleteHandler();
+              case "OPTIONS":
+                  return new OptionsHandler(router);
+              default:
+                  return new ErrorHandler(Status.NotFound);
+            }
         }
+        return new ErrorHandler(Status.NotFound);
     }
 }
